requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pandas
numpy
pytest
pytest-asyncio
httpx
python-dotenv
pyyaml
email-validator
